requirements.txt
pywebview
opencv-python  
numpy  
pillow  
tqdm  
pyyaml  
scikit-image  
scipy  